refactor: Extract CELVQNGMixin to deduplicate CELVQ-NG family
Move shared gamma scheduling, NG rank-weighted pooling, and
cross-entropy loss into a mixin class. Each model now only
implements _compute_distances() for its metric-specific logic.

Reduces ~1050 lines to ~530 lines with zero behavioral change.
All 60 CELVQ-NG tests pass, all examples produce identical results.
